Optimization ffmpeg

+++ camworker.cpp
@@ -20,6 +20,7 @@ CamWorker::CamWorker(const QString& camId, const Config& config, QObject* parent
     dstIp = config.ip;
 
     dstPort = config.port;
+    metaPort = dstPort + 100;
 
     timeInterval = config.timeInterval;
 
@@ -28,8 +29,30 @@ CamWorker::CamWorker(const QString& camId, const Config& config, QObject* parent
     mode = config.mode;
     width = config.width;
     height = config.height;
+    rawSize = config.rawSize;
 
     frames = timeInterval / 100;
+
+
+    ffmpegUrl = QString("tcp://%1:%2").arg(dstIp).arg(dstPort);
+    ffmpegArgs << "-f" << "rawvideo"
+            << "-loglevel" << "debug"
+            << "-pixel_format" << "bayer_rggb8"
+            << "-video_size" << QString("%1x%2").arg(width).arg(height)
+            << "-framerate" << "10"
+            << "-i" << "pipe:0"
+            << "-vf" << "format=bgr24"
+            << "-c:v" << "libx264"
+            << "-preset" << "veryfast"
+            << "-tune" << "zerolatency"
+            << "-pix_fmt" << "yuv420p"
+            << "-f" << "mpegts"
+            << ffmpegUrl;
+
+    frameBuffer.resize(static_cast<int>(static_cast<qint64>(width) * height));
+    ffmpeg.setProgram("ffmpeg");
+    ffmpeg.setArguments(ffmpegArgs);
+    ffmpeg.setProcessChannelMode(QProcess::SeparateChannels);
 
     connect(&watcher, &QFileSystemWatcher::directoryChanged, this, &CamWorker::onFileSystemChanged);
 
@@ -45,14 +68,56 @@ CamWorker::CamWorker(const QString& camId, const Config& config, QObject* parent
 
 CamWorker::~CamWorker()
 {
-    socket->flush();
-    socket->disconnectFromHost();
-    socket->waitForDisconnected(3000);
+    stopFfmpeg(false);
+
+    if(socket->state() == QAbstractSocket::ConnectedState)
+    {
+        socket->flush();
+        socket->disconnectFromHost();
+        socket->waitForDisconnected(3000);
+    }
 
     socket->deleteLater();
     timer.deleteLater();
 }
 
+bool CamWorker::ensureMetaSocketConnected()
+{
+    if(socket->state() == QAbstractSocket::ConnectedState)
+        return true;
+
+    if(socket->state() != QAbstractSocket::UnconnectedState)
+        socket->abort();
+
+    socket->connectToHost(dstIp, metaPort);
+
+    if(!socket->waitForConnected(3000))
+    {
+        QString errString = QString("fail connect to ip : %1 port : %2").arg(dstIp).arg(metaPort);
+        qCritical() << errString;
+        return false;
+    }
+
+    return true;
+}
+
+bool CamWorker::sendMetaHeader(const QByteArray& header)
+{
+    if(!ensureMetaSocketConnected())
+        return false;
+
+    socket->write(header);
+
+    if(!socket->waitForBytesWritten(3000))
+    {
+        qCritical() << "write fail: " << socket->errorString();
+        socket->abort();
+        return false;
+    }
+
+    socket->flush();
+    return true;
+}
 
 void CamWorker::init()
 {
@@ -115,16 +180,100 @@ void CamWorker::onFileSystemChanged(const QString& path)
                                            QDir::Files | QDir::NoDotAndDotDot,
                                            QDir::Name);
 
-        if(rawFiles.size() == 3000)
+        if(rawFiles.size() == rawSize)
             sensorDirs.enqueue(path);
     }
+}
+
+bool CamWorker::startFfmpeg()
+{
+    if(ffmpeg.state() != QProcess::NotRunning)
+        return true;
+
+    ffmpeg.start();
+
+    if(!ffmpeg.waitForStarted(3000))
+    {
+        qCritical() << "ffmpeg start fail: " << ffmpeg.errorString();
+        return false;
+    }
+
+    return true;
+}
+
+void CamWorker::stopFfmpeg(bool forceKill)
+{
+    if(ffmpeg.state() == QProcess::NotRunning)
+        return;
+
+    if(forceKill)
+    {
+        ffmpeg.kill();
+        ffmpeg.waitForFinished();
+        return;
+    }
+
+    ffmpeg.closeWriteChannel();
+    qCritical() << "ffmpeg failed to wait for termination";
+    ffmpeg.kill();
+    ffmpeg.waitForFinished();
+}
+
+bool CamWorker::writeFrameToFfmpeg(const QString& filePath)
+{
+    QFile in(filePath);
+    if(!in.open(QIODevice::ReadOnly))
+    {
+        qCritical() << "Fail to open raw file: " << filePath;
+        return false;
+    }
+
+    const qint64 rawBytes = frameBuffer.size();
+    qint64 readBytes = in.read(frameBuffer.data(), rawBytes);
+    in.close();
+
+    if(readBytes != rawBytes)
+    {
+        qCritical() << "Fail to read raw file : "
+                    << filePath
+                    << "read = " << readBytes
+                    << "expected = " << rawBytes;
+
+        return false;
+    }
+
+    qint64 written = 0;
+
+    while(written < frameBuffer.size())
+    {
+        qint64 chunk = ffmpeg.write(frameBuffer.constData() + written, frameBuffer.size() - written);
+        if(chunk < 0)
+        {
+            qCritical() << "fail to write to ffmpeg stdin" << filePath;
+            return false;
+        }
+
+        written += chunk;
+        if(!ffmpeg.waitForBytesWritten(-1))
+        {
+            qCritical() << "ffmpeg stdin flush 실패:" << filePath;
+            qCritical() << "state =" << ffmpeg.state();
+            qCritical() << "exitCode =" << ffmpeg.exitCode();
+            qCritical() << "error =" << ffmpeg.errorString();
+            qCritical().noquote() << "stderr:\n" << ffmpeg.readAllStandardError();
+            return false;
+        }
+    }
+
+    return true;
+
 }
 
 
 void CamWorker::start()
 {
     qDebug() << "Start to vss";
-    if(sensorDirs.isEmpty())
+    if(sensorDirs.isEmpty() && rawFiles.isEmpty())
     {
         qDebug() << "No sensorDirs";
         return;
@@ -156,20 +305,6 @@ void CamWorker::start()
     requestCreateClip();
 }
 
-void CamWorker::sensorMode()
-{
-    qDebug() << "Sensor Dir list size " << sensorDirs.size();
-//    QVector<QString> clips;
-//    QString camPath = QString("%1/%2").arg(currDir).arg(id);
-//    QDir dir(camPath);
-
-//    while((dir.entryList()).size() != 3000)
-//    {
-
-//    }
-
-//    sendClip(clips);
-}
 
 void CamWorker::requestCreateClip()
 {
@@ -195,20 +330,16 @@ void CamWorker::requestCreateClip()
 
 void CamWorker::sendClip(const QVector<QString>& clips)
 {
+    if(clips.isEmpty())
+    {
+        qDebug() << "No clips to send";
+        return;
+    }
+
     QString timestamp = QDateTime::currentDateTime().toString("yyyyMMdd_hhmmss_zzz");
 
 
-    int metaPort = dstPort + 100;
-    socket->connectToHost(dstIp, metaPort);
-
-    if(!socket->waitForConnected(3000))
-    {
-        QString errString = QString("fail connect to ip : %1 port : %2").arg(dstIp).arg(metaPort);
-        qCritical() << errString;
-    }
-
     QJsonObject obj;
-//    obj["folder"] = currDir;
 
     QFileInfo fi(currDir);
     QDir parentDir = fi.dir();
@@ -219,129 +350,33 @@ void CamWorker::sendClip(const QVector<QString>& clips)
     QByteArray header = QJsonDocument(obj).toJson(QJsonDocument::Compact);
     header.append("\n");
 
-    socket->write(header);
-
-    if (!socket->waitForBytesWritten(3000)) {
-        qCritical() << "write fail:" << socket->errorString();
-    }
-
-    socket->flush();
+    if(!sendMetaHeader(header))
+        return;
 
     qDebug() << "Success to send meta: " << header;
 
-    QProcess ffmpeg;
+    QElapsedTimer eTimer;
+    eTimer.start();
 
-    QString url = QString("tcp://%1:%2").arg(dstIp).arg(dstPort);
-
-    QStringList args = {
-            "-f", "rawvideo",
-            "-loglevel", "debug",
-            "-pixel_format", "bayer_rggb8",
-            "-video_size", QString("%1x%2").arg(width).arg(height),
-            "-framerate", "10",
-            "-i", "pipe:0",
-            "-vf", "format=bgr24",
-            "-c:v", "libx264",
-            "-preset", "veryfast",
-            "-tune", "zerolatency",
-            "-pix_fmt", "yuv420p",
-            "-f", "mpegts",
-            url
-        };
-
-    ffmpeg.setProgram("ffmpeg");
-    ffmpeg.setArguments(args);
-    ffmpeg.setProcessChannelMode(QProcess::SeparateChannels);
-
-    QElapsedTimer timer;
-    timer.start();
-
-    ffmpeg.start("ffmpeg", args);
-
-    if (!ffmpeg.waitForStarted(3000))
-    {
-        qCritical() << "ffmpeg start fail";
+    if(!startFfmpeg())
         return;
-    }
 
-    qint64 rawBytes = (qint64)width * height;
-
-    for (const QString& file : clips)
+    for(const QString& file : clips)
     {
-        QFile in(file);
-        if (!in.open(QIODevice::ReadOnly)) continue;
-
-        QByteArray frame(rawBytes, Qt::Uninitialized);
-        qint64 readBytes = in.read(frame.data(), rawBytes);
-
-        in.close();
-
-        if(readBytes != rawBytes)
+        if(!writeFrameToFfmpeg(file))
         {
-            qCritical() << "Fail to read raw file : "
-                        << in.fileName()
-                        << "read = " << readBytes
-                        << "expected = " << rawBytes;
-
-            // TODO
-            // handle error
-            ffmpeg.kill();
-            ffmpeg.waitForFinished();
+            stopFfmpeg(true);
             return;
         }
-
-
-        qint64 written = 0;
-        while(written < frame.size())
-        {
-            qint64 chunk = ffmpeg.write(frame.constData() + written, frame.size() - written);
-            if(chunk < 0)
-            {
-                qCritical() << "fail to write to ffmpeg stdin" << in.fileName();
-                ffmpeg.kill();
-                ffmpeg.waitForFinished();
-                return;
-            }
-
-            written += chunk;
-            if(!ffmpeg.waitForBytesWritten(-1))
-            {
-                qCritical() << "ffmpeg stdin flush 실패:" << in.fileName();
-                qCritical() << "state =" << ffmpeg.state();
-                qCritical() << "exitCode =" << ffmpeg.exitCode();
-                qCritical() << "error =" << ffmpeg.errorString();
-                qCritical().noquote() << "stderr:\n" << ffmpeg.readAllStandardError();
-                ffmpeg.kill();
-                ffmpeg.waitForFinished();
-                return;
-            }
-        }
     }
 
-    ffmpeg.closeWriteChannel();
+    stopFfmpeg(false);
 
-    if(!ffmpeg.waitForFinished(-1))
-    {
-        qCritical() << "ffmpeg failed to wait for termination";
-        return;
-    }
-
-    qint64 elapsedMs = timer.elapsed();
-    qDebug() << "ffmpeg encoding time(ms):" << elapsedMs;
-
-    QByteArray stdOut = ffmpeg.readAllStandardOutput();
-    QByteArray stdErr = ffmpeg.readAllStandardError();
-
-
-//    if (!stdOut.isEmpty()) {
-//        qInfo().noquote() << "ffmpeg stdout:\n" << QString::fromLocal8Bit(stdOut);
-//    }
-//    if (!stdErr.isEmpty()) {
-//        qInfo().noquote() << "ffmpeg stderr:\n" << QString::fromLocal8Bit(stdErr);
-//    }
+    qint64 elapsedMs = eTimer.elapsed();
+    qDebug() << "ffmpeg encoding time(ms): " << elapsedMs;
 
     if (ffmpeg.exitCode() != 0) {
-        qCritical() << "ffmpeg 인코딩 실패. exitCode =" << ffmpeg.exitCode();
+        qCritical() << "ffmpeg fail to encoding. exitCode =" << ffmpeg.exitCode();
         return;
     }
 
@@ -358,9 +393,8 @@ void CamWorker::getAnswer(QByteArray data)
     QJsonDocument doc = QJsonDocument::fromJson(data, &parseError);
 
     if(parseError.error != QJsonParseError::NoError)
-    {
         qCritical() << "JSON parsing error: " << parseError.errorString();
-    }
+
 
     QJsonObject root = doc.object();
 
@@ -368,32 +402,41 @@ void CamWorker::getAnswer(QByteArray data)
     QString answer = root.value("answer").toString();
 
     QString path = rootPath + "/" + folderName;
-    if(answer.contains("yes", Qt::CaseInsensitive) || answer.contains("snow", Qt::CaseInsensitive) || answer.contains("rain", Qt::CaseInsensitive))
+    if(answer.contains("yes", Qt::CaseInsensitive))
     {
-        trashList.remove(path);
+        if(trashList.contains(path))
+        {
+            qDebug() << "remove folder from trash queue" << path;
+            trashList.remove(path);
+        }
     }
     else
     {
         QString path = rootPath + "/" + folderName;
-        trashList.insert(path);
+        qDebug() << "insert folder to trash queue" << path;
+        if(!trashList.contains(path))
+            trashList.insert(path);
     }
 
     QFileInfo fi(currDir);
     QDir parentDir = fi.dir();
 
-
-    for(const QString& p : trashList)
+    if(!trashList.isEmpty())
     {
-        if(p == parentDir.absolutePath())
-            continue;
-
-        QDir removeDir(p);
-        if(removeDir.exists())
+        for(const QString& p : trashList)
         {
-            if(removeDir.removeRecursively())
-                qDebug() << "Success to remove folder : " << p;
-            else
-                qDebug() << "Failed to remove folder : " << p;
+            if(p == parentDir.absolutePath())
+                continue;
+
+            QDir removeDir(p);
+            if(removeDir.exists())
+            {
+                if(removeDir.removeRecursively())
+                    qDebug() << "Success to remove folder : " << p;
+                else
+                    qDebug() << "Failed to remove folder : " << p;
+            }
         }
     }
+
 }
